+++ index.ts
@@ -2,6 +2,7 @@ export * from './src/bindings';
 export * from './src/check';
 export * from './src/engineGlobals';
 export * from './src/fieldMap';
+export * from './src/fuzzy';
 export * from './src/lens';
 export * from './src/operator';
 export * from './src/operatorCatalog';